@@ -18,45 +18,44 @@
     "postinstall": "prisma generate"
   },
   "dependencies": {
-    "@envelop/auth0": "4.0.6",
-    "@envelop/core": "3.0.6",
-    "@envelop/depth-limit": "2.0.6",
-    "@envelop/generic-auth": "5.0.6",
-    "@envelop/rate-limiter": "4.0.6",
-    "@fastify/compress": "6.2.1",
-    "@fastify/cors": "8.2.1",
-    "@fastify/helmet": "^9.1.0",
-    "@graphql-tools/schema": "9.0.18",
-    "@graphql-tools/utils": "^9.1.4",
-    "@prisma/client": "4.8.1",
+    "@envelop/auth0": "6.0.0",
+    "@envelop/core": "5.0.0",
+    "@envelop/depth-limit": "4.0.0",
+    "@envelop/generic-auth": "7.0.0",
+    "@fastify/compress": "7.0.0",
+    "@fastify/cors": "9.0.1",
+    "@fastify/helmet": "^11.1.1",
+    "@graphql-tools/schema": "10.0.3",
+    "@graphql-tools/utils": "^10.1.0",
+    "@prisma/client": "5.10.2",
     "@types/got": "9.6.12",
-    "@types/jsonwebtoken": "8.5.9",
-    "@types/node": "16.18.24",
-    "fastify": "4.13.0",
-    "got": "11.8.6",
-    "graphql": "16.6.0",
-    "graphql-tools": "8.3.20",
-    "graphql-yoga": "3.9.1",
-    "jsonwebtoken": "8.5.1",
-    "jwks-rsa": "2.1.5",
-    "prisma": "4.8.1",
-    "ts-pattern": "^4.0.5",
+    "@types/jsonwebtoken": "9.0.6",
+    "@types/node": "20.11.24",
+    "fastify": "4.26.2",
+    "got": "14.2.0",
+    "graphql": "16.8.1",
+    "graphql-tools": "9.0.1",
+    "graphql-yoga": "5.1.1",
+    "jsonwebtoken": "9.0.2",
+    "jwks-rsa": "3.1.0",
+    "prisma": "5.10.2",
+    "ts-pattern": "^5.0.8",
     "tsconfig": "workspace:*",
-    "typescript": "4.9.5",
+    "typescript": "5.3.3",
     "validation-schema": "workspace:*",
-    "zod": "3.20.6"
+    "zod": "3.22.4"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "2.16.5",
-    "@graphql-codegen/typescript-resolvers": "2.7.13",
-    "@graphql-codegen/visitor-plugin-common": "2.13.8",
-    "@graphql-yoga/render-graphiql": "3.6.1",
-    "chalk": "4.1.2",
+    "@graphql-codegen/cli": "5.0.2",
+    "@graphql-codegen/typescript-resolvers": "4.0.6",
+    "@graphql-codegen/visitor-plugin-common": "5.1.0",
+    "@graphql-yoga/render-graphiql": "5.1.1",
+    "chalk": "5.3.0",
     "config": "workspace:*",
-    "dotenv-cli": "6.0.0",
-    "nodemon": "2.0.22",
-    "tsc-alias": "1.8.5",
-    "tsconfig-paths": "4.1.2",
-    "tsx": "3.12.6"
+    "dotenv-cli": "7.3.0",
+    "nodemon": "3.1.0",
+    "tsc-alias": "1.8.8",
+    "tsconfig-paths": "4.2.0",
+    "tsx": "4.7.1"
   }
 }
