@@ -10,69 +10,64 @@
     "type-check": "tsc --noEmit",
     "codegen:grqphql": "graphql-codegen --config codegen.yml",
     "codegen:watch": "graphql-codegen --watch --config codegen.yml",
-    "test": "NODE_ENV=test jest --config ./jest.config.js",
-    "test:watch": "NODE_ENV=test jest --watch --config ./jest.config.js",
+    "test": "SET NODE_ENV=test jest --config ./jest.config.js",
+    "test:watch": "SET NODE_ENV=test jest --watch --config ./jest.config.js",
     "bundle-analyzer": "ANALYZE=true pnpm run build"
   },
   "dependencies": {
-    "@auth0/auth0-react": "1.12.1",
-    "@heroicons/react": "1.0.6",
-    "@hookform/resolvers": "^2.9.10",
-    "@next/font": "^13.1.6",
-    "@radix-ui/colors": "0.1.8",
-    "@radix-ui/react-avatar": "1.0.2",
-    "@radix-ui/react-dialog": "1.0.3",
+    "@auth0/auth0-react": "2.2.4",
+    "@heroicons/react": "2.1.1",
+    "@hookform/resolvers": "^3.3.4",
+    "@radix-ui/colors": "3.0.0",
+    "@radix-ui/react-avatar": "1.0.4",
+    "@radix-ui/react-dialog": "1.0.5",
     "@radix-ui/react-icons": "1.3.0",
-    "@radix-ui/react-portal": "1.0.2",
-    "autoprefixer": "^10.4.13",
-    "graphql": "16.6.0",
-    "graphql-anywhere": "4.2.8",
-    "next": "13.3.1",
-    "postcss": "^8.4.21",
+    "@radix-ui/react-portal": "1.0.4",
+    "autoprefixer": "^10.4.18",
+    "graphql": "16.8.1",
+    "next": "14.1.2",
+    "postcss": "^8.4.35",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "7.43.9",
+    "react-hook-form": "7.51.0",
     "spinners-react": "1.0.7",
-    "tailwindcss": "3.2.7",
-    "ts-pattern": "3.3.5",
-    "typescript": "4.9.5",
+    "tailwindcss": "3.4.1",
+    "ts-pattern": "5.0.8",
+    "typescript": "5.3.3",
     "ui": "workspace:*",
-    "urql": "2.2.3",
+    "urql": "4.0.6",
     "validation-schema": "workspace:*",
-    "wonka": "4.0.15",
-    "zod": "3.20.6"
+    "wonka": "6.3.4",
+    "zod": "3.22.4"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "2.16.5",
-    "@graphql-codegen/gql-tag-operations-preset": "1.7.4",
-    "@graphql-codegen/schema-ast": "2.6.1",
-    "@graphql-codegen/typed-document-node": "2.3.13",
-    "@graphql-codegen/typescript": "2.8.8",
-    "@graphql-codegen/typescript-operations": "2.5.13",
-    "@graphql-codegen/typescript-urql": "3.7.3",
-    "@graphql-typed-document-node/core": "3.1.2",
-    "@next/bundle-analyzer": "12.3.4",
-    "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "13.4.0",
-    "@testing-library/user-event": "14.4.3",
-    "@types/jest": "27.5.2",
-    "@types/node": "16.18.24",
-    "@types/react": "18.0.38",
-    "@types/react-dom": "18.0.11",
-    "@types/styled-components": "5.1.26",
-    "@types/testing-library__jest-dom": "^5.14.5",
-    "babel-jest": "27.5.1",
+    "@graphql-codegen/cli": "5.0.2",
+    "@graphql-codegen/gql-tag-operations-preset": "2.1.0",
+    "@graphql-codegen/schema-ast": "4.0.2",
+    "@graphql-codegen/typed-document-node": "5.0.6",
+    "@graphql-codegen/typescript": "4.0.6",
+    "@graphql-codegen/typescript-operations": "4.2.0",
+    "@graphql-codegen/typescript-urql": "4.0.0",
+    "@graphql-typed-document-node/core": "3.2.0",
+    "@next/bundle-analyzer": "14.1.2",
+    "@testing-library/jest-dom": "6.4.2",
+    "@testing-library/react": "14.2.1",
+    "@testing-library/user-event": "14.5.2",
+    "@types/jest": "29.5.12",
+    "@types/node": "20.11.24",
+    "@types/react": "18.2.63",
+    "@types/react-dom": "18.2.20",
+    "@types/styled-components": "5.1.34",
     "config": "workspace:*",
     "graphql-playground-html": "1.6.30",
-    "graphql-typescript-integration": "1.2.1",
-    "isomorphic-unfetch": "3.1.0",
-    "jest": "27.5.1",
-    "msw": "0.44.2",
-    "prettier-plugin-tailwindcss": "^0.2.2",
+    "isomorphic-unfetch": "4.0.2",
+    "jest": "29.7.0",
+    "msw": "2.2.2",
+    "prettier-plugin-tailwindcss": "^0.5.11",
     "react-is": "18.2.0",
     "react-test-renderer": "18.2.0",
     "tailwind-config": "workspace:*",
-    "ts-jest": "27.1.5",
+    "ts-jest": "29.1.2",
     "tsconfig": "workspace:*"
   }
 }
