@@ -2,7 +2,8 @@
   "name": "ui",
   "version": "1.0.0",
   "description": "",
-  "types": "./dist/index.d.ts",
+  "main": "dist/index.mjs",
+  "types": "./dist/index.d.mts",
   "exports": {
     ".": "./dist",
     "./styles.css": "./src/styles.css"
@@ -24,24 +25,24 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@radix-ui/react-avatar": "1.0.2",
-    "@radix-ui/react-dialog": "1.0.3",
-    "@radix-ui/react-portal": "1.0.2",
-    "@types/node": "16.18.24",
-    "@types/react": "18.0.38",
-    "@types/react-dom": "18.0.11",
-    "autoprefixer": "^10.4.13",
-    "class-variance-authority": "^0.4.0",
-    "clsx": "^1.2.1",
-    "lucide-react": "^0.161.0",
-    "postcss": "^8.4.21",
+    "@radix-ui/react-avatar": "1.0.4",
+    "@radix-ui/react-dialog": "1.0.5",
+    "@radix-ui/react-portal": "1.0.4",
+    "@types/node": "20.11.24",
+    "@types/react": "18.2.63",
+    "@types/react-dom": "18.2.20",
+    "autoprefixer": "^10.4.18",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.0",
+    "lucide-react": "^0.344.0",
+    "postcss": "^8.4.35",
     "spinners-react": "1.0.7",
     "tailwind-config": "workspace:*",
-    "tailwind-merge": "^1.9.0",
-    "tailwindcss": "3.2.7",
-    "tailwindcss-animate": "^1.0.5",
+    "tailwind-merge": "^2.2.1",
+    "tailwindcss": "3.4.1",
+    "tailwindcss-animate": "^1.0.7",
     "tsconfig": "workspace:*",
-    "tsup": "^6.6.0",
-    "typescript": "4.9.5"
+    "tsup": "^8.0.2",
+    "typescript": "5.3.3"
   }
 }
