@@ -2,8 +2,8 @@
   "name": "validation-schema",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/index",
-  "types": "dist/index.d.ts",
+  "main": "dist/index.mjs",
+  "types": "dist/index.d.mts",
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
@@ -14,14 +14,14 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@graphql-codegen/cli": "3.3.1",
-    "@graphql-codegen/typescript": "3.0.4",
-    "@types/node": "16.18.24",
-    "graphql": "16.6.0",
-    "graphql-codegen-typescript-validation-schema": "^0.7.1",
+    "@graphql-codegen/cli": "5.0.2",
+    "@graphql-codegen/typescript": "4.0.6",
+    "@types/node": "20.11.24",
+    "graphql": "16.8.1",
+    "graphql-codegen-typescript-validation-schema": "^0.12.1",
     "tsconfig": "workspace:*",
-    "tsup": "^6.6.0",
-    "typescript": "4.9.5",
-    "zod": "3.20.6"
+    "tsup": "^8.0.2",
+    "typescript": "5.3.3",
+    "zod": "3.22.4"
   }
 }
