@@ -1,4 +1,4 @@
-import { graphql } from 'msw'
+import { graphql, HttpResponse } from 'msw'
 
 import { GetTodosQuery, GetTodosQueryVariables } from '~/generated/graphql'
 
@@ -7,11 +7,12 @@ import { todoFactory } from './factories/todo'
 export const handlers = [
   graphql.query<GetTodosQuery, GetTodosQueryVariables>(
     'GetTodos',
-    (req, res, ctx) =>
-      res(
-        ctx.data({
-          todosByCurrentUser: [todoFactory(), todoFactory()],
-        })
-      )
+    () => {
+      return HttpResponse.json({
+        data: {
+          todosByCurrentUser: [todoFactory(), todoFactory()]
+        }
+      })
+    }
   ),
 ]
