@@ -2,8 +2,8 @@ import '../styles/globals.css'
 import 'ui/styles.css'
 
 import { Auth0Provider } from '@auth0/auth0-react'
-import { Inter } from '@next/font/google'
 import { AppProps } from 'next/app'
+import { Inter } from 'next/font/google'
 
 import { Navbar, UrqlClientProvider } from '~/components/'
 import { CurrentUserProvider } from '~/contexts/currentUser'
@@ -25,10 +25,10 @@ function MyApp({ Component, pageProps }: AppProps) {
       <Auth0Provider
         domain={process.env.NEXT_PUBLIC_AUTH0_DOMAIN || ''}
         clientId={process.env.NEXT_PUBLIC_AUTH0_CLIENT_ID || ''}
-        audience={process.env.NEXT_PUBLIC_AUTH0_AUDIENCE || ''}
-        redirectUri={
-          typeof window !== 'undefined' ? window.location.origin : undefined
-        }
+        authorizationParams={{
+          audience: process.env.NEXT_PUBLIC_AUTH0_AUDIENCE || '',
+          redirect_uri: typeof window !== 'undefined' ? window.location.origin : undefined,
+        }}
       >
         <UrqlClientProvider>
           <CurrentUserProvider>
