@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { GraphQLHandler, GraphQLRequest } from 'msw'
+import { GraphQLHandler } from 'msw'
 import React, { StrictMode } from 'react'
 
 import { UrqlClientProvider } from '~/components/util/UrqlClientProvider'
@@ -9,17 +9,17 @@ import { server } from '~/mocks/server'
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export const testRenderer =
   (children: React.ReactNode) =>
-  (responseOverride?: GraphQLHandler<GraphQLRequest<never>>) => {
-    if (responseOverride) {
-      server.use(responseOverride)
-    }
-    render(
-      <StrictMode>
-        <UrqlClientProvider>{children}</UrqlClientProvider>
-      </StrictMode>
-    )
+    (responseOverride?: GraphQLHandler) => {
+      if (responseOverride) {
+        server.use(responseOverride)
+      }
+      render(
+        <StrictMode>
+          <UrqlClientProvider>{children}</UrqlClientProvider>
+        </StrictMode>
+      )
 
-    return {
-      user: userEvent.setup(),
+      return {
+        user: userEvent.setup(),
+      }
     }
-  }
